requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
python-dateutil==2.8.2
selenium==4.10.0
undetected-chromedriver==3.5.5
selenium-wire==5.1.0
seleniumbase==4.28.0
cloudscraper==1.2.71
PyPDF2==3.0.1
anthropic==0.25.0
# BlueSky client; newer versions fix API model mismatches
atproto>=0.0.49,<1.0
